fix: make every type float in midnight
Making everything of type double is causing a lot of incompatibilities with OpenGL, so they're now floats.

Some post-testing might need to be done.

+++ lib/midnight/matrix.cxx
@@ -4,12 +4,12 @@ namespace midnight
 {
 	Matrix<3, 1> cross(const Matrix<3, 1> v1, const Matrix<3, 1> v2)
 	{
-		const double x1{v1.entry(0, 0)};
-		const double y1{v1.entry(1, 0)};
-		const double z1{v1.entry(2, 0)};
-		const double x2{v2.entry(0, 0)};
-		const double y2{v2.entry(1, 0)};
-		const double z2{v2.entry(2, 0)};
+		const float x1{v1.entry(0, 0)};
+		const float y1{v1.entry(1, 0)};
+		const float z1{v1.entry(2, 0)};
+		const float x2{v2.entry(0, 0)};
+		const float y2{v2.entry(1, 0)};
+		const float z2{v2.entry(2, 0)};
 		return Matrix<3, 1>{
 			y1 * z2 - y2 * z1,
 			z1 * x2 - z2 * x1,
@@ -17,56 +17,56 @@ namespace midnight
 		};
 	}
 
-	Matrix<4, 4> matrixRotation(const Matrix<3, 1> line, const double angle)
+	Matrix<4, 4> matrixRotation(const Matrix<3, 1> line, const float angle)
 	{
 		using std::sin, std::cos, std::pow;
-		const double r{angle};
+		const float r{angle};
 		const Matrix<3, 1> l(normalise(line));
-		const double xl{l.entry(0, 0)};
-		const double yl{l.entry(1, 0)};
-		const double zl{l.entry(2, 0)};
+		const float xl{l.entry(0, 0)};
+		const float yl{l.entry(1, 0)};
+		const float zl{l.entry(2, 0)};
 		return Matrix<4, 4>{
-			(pow(xl, 2) * (1 - cos(r))) + cos(r),
+			(pow(xl, 2.0F) * (1 - cos(r))) + cos(r),
 			(xl * yl) * (1 - cos(r)) + (zl * sin(r)),
 			(xl * zl) * (1 - cos(r)) - (yl * sin(r)),
 			0,
 
 			(xl * yl) * (1 - cos(r)) - (zl * sin(r)),
-			(pow(yl, 2) * (1 - cos(r))) + cos(r),
+			(pow(yl, 2.0F) * (1 - cos(r))) + cos(r),
 			(yl * zl) * (1 - cos(r)) + (xl * sin(r)),
 			0,
 
 			(xl * zl) * (1 - cos(r)) + (yl * sin(r)),
 			(yl * zl) * (1 - cos(r)) - (xl * sin(r)),
-			(pow(zl, 2) * (1 - cos(r))) + cos(r),
+			(pow(zl, 2.0F) * (1 - cos(r))) + cos(r),
 			0,
 
 			0, 0, 0, 1
 		};
 	}
 
-	Matrix<4, 4> matrixScale(const Matrix<3, 1> line, const double factor)
+	Matrix<4, 4> matrixScale(const Matrix<3, 1> line, const float factor)
 	{
 		using std::pow;
 		const Matrix<3, 1> l(normalise(line));
-		const double xl{l.entry(0, 0)};
-		const double yl{l.entry(1, 0)};
-		const double zl{l.entry(2, 0)};
-		const double f{factor};
+		const float xl{l.entry(0, 0)};
+		const float yl{l.entry(1, 0)};
+		const float zl{l.entry(2, 0)};
+		const float f{factor};
 		return Matrix<4, 4>{
-			pow(xl, 2) * (f - 1) + 1,
+			pow(xl, 2.0F) * (f - 1) + 1,
 			xl * yl * (f - 1),
 			xl * zl * (f - 1),
 			0,
 
 			xl * yl * (f - 1),
-			pow(yl, 2) * (f - 1) + 1,
+			pow(yl, 2.0F) * (f - 1) + 1,
 			yl * zl * (f - 1),
 			0,
 
 			xl * zl * (f - 1),
 			yl * zl * (f - 1),
-			pow(zl, 2) * (f - 1) + 1,
+			pow(zl, 2.0F) * (f - 1) + 1,
 
 			0, 0, 0, 1
 		};
@@ -82,10 +82,10 @@ namespace midnight
 		};
 	}
 
-	Matrix<4, 4> matrixPerspective(const double fovx, const double aspect, const double near, const double far)
+	Matrix<4, 4> matrixPerspective(const float fovx, const float aspect, const float near, const float far)
 	{
-		const double zoomx{1 / (std::tan(fovx / 2))};
-		const double zoomy{aspect * zoomx};
+		const float zoomx{1 / (std::tan(fovx / 2))};
+		const float zoomy{aspect * zoomx};
 		return Matrix<4, 4>{
 			zoomx, 0, 0, 0,
 			0, zoomy, 0, 0,
@@ -94,7 +94,7 @@ namespace midnight
 		};
 	}
 
-	Matrix<4, 4> matrixOrthographic(const double width, const double height, const double near, const double far)
+	Matrix<4, 4> matrixOrthographic(const float width, const float height, const float near, const float far)
 	{
 		return Matrix<4, 4>{
 			2 / width, 0, 0, 0,
